c_dumper: implement strong/weak edge ordering and fix smaller issues
This commit adds a lot of test cases, revamps the way
struct/union/typedef/enum/fwd order is done to consider "strong" vs
"weak" edges between BTF types. Strong edge is created between
struct/union and entities (typedefs, enum, other struct/union, fwd) that
are embedded as part of a struct. "Weak" are those that can be
partially-defined through forward declarations because they are
referenced through path containing a pointer (e.g, pointer to a struct
or func_proto taking/returning a struct).

There is a subtlety there, as in some cases struct/union definition has
to be inlined (for anonymous structs), which automatically makes edge
"strong", because that struct and it's embedded members can't be just
forward-declared.

+++ src/btf.rs
@@ -485,8 +485,8 @@ impl Btf {
         &self.types[type_id as usize]
     }
 
-    pub fn type_cnt(&self) -> usize {
-        self.types.len()
+    pub fn type_cnt(&self) -> u32 {
+        self.types.len() as u32
     }
 
     pub fn load<'data>(elf: object::ElfFile<'data>) -> BtfResult<Btf> {
